requirements.txt
pytest==7.4.3
requests==2.31.0
flask==3.0.0
sqlalchemy==2.0.23
pytest-flask==1.3.0
python-dotenv==1.0.0
bcrypt==4.1.1
pyjwt==2.8.0
pandas==2.1.3
numpy==1.25.2 